Fix slider sending double values in API payload

+++ frontend/lib/preferences/user_preferences.dart
@@ -7,7 +7,7 @@ class UserPreferences {
     required this.dietPreferences,
   });
 
-  final double distanceMeters;
+  final int distanceMeters;
   final int budgetLevel;
   final Set<DietPreference> dietPreferences;
